bankey/pyproject.toml
[tool.poetry]
name = "bankey"
version = "0.1.0"
description = "Simple Banking App"
authors = ["Aleksandra Zwierzyńska"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
black = "^22.8.0"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]
black = "^22.8.0"
pytest = "^7.1.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
